+++ src/mobileComponents/sections/Certifications/components/CertCard.mobile.mobile.jsx
