+++ src/mobileComponents/sections/Experience/CaseStudyModal.mobile.mobile.jsx
