+++ src/mobileComponents/sections/Goals/components/SectionHeader.mobile.mobile.jsx
