+++ src/mobileComponents/sections/Goals/components/TimelineItem.mobile.mobile.jsx
