+++ src/mobileComponents/sections/Skills/components/SkillCard.mobile.mobile.jsx
