+++ src/mobileComponents/sections/Skills/components/SkillCategory.mobile.mobile.jsx
